Fix compile for armv7 as per https://github.com/golang/go/issues/23086#issuecomment-371017565

+++ commonxl/cell.go
@@ -217,7 +217,7 @@ func NewCell(value interface{}) Cell {
 		c[1] = IntegerCell
 
 	case uint:
-		if v > math.MaxInt64 {
+		if int64(v) > int64(math.MaxInt64) {
 			c[0] = float64(v)
 			c[1] = FloatCell
 		} else {
